+++ test/utils.py
@@ -20,6 +20,7 @@ from qiskit.compiler import assemble, transpile
 from qiskit.test.reference_circuits import ReferenceCircuits
 from qiskit.providers.exceptions import JobError
 from qiskit.providers.jobstatus import JobStatus
+from qiskit.providers.ibmq.ibmqfactory import IBMQFactory
 from qiskit.providers.ibmq.accountprovider import AccountProvider
 from qiskit.providers.ibmq.ibmqbackend import IBMQBackend
 from qiskit.providers.ibmq.job import IBMQJob
@@ -102,3 +103,34 @@ def cancel_job(job: IBMQJob, verify: bool = False) -> bool:
             pass
 
     return cancelled
+
+
+def get_provider(
+        ibmq_factory: IBMQFactory,
+        qe_token: str,
+        qe_url: str,
+        default: bool = True
+) -> AccountProvider:
+    """Return a provider for the account.
+
+    Args:
+        ibmq_factory: An `IBMQFactory` instance.
+        qe_token: IBM Quantum Experience token.
+        qe_url: IBM Quantum Experience auth URL.
+        default: If `True`, the default open access project provider is returned.
+            Otherwise, a non open access project provider is returned.
+
+    Returns:
+        A provider, as specified by `default`.
+    """
+    provider_to_return = ibmq_factory.enable_account(qe_token, url=qe_url)  # Default provider.
+    if not default:
+        # Get a non default provider (i.e.not the default open access project).
+        providers = ibmq_factory.providers()
+        for provider in providers:
+            if provider != provider_to_return:
+                provider_to_return = provider
+                break
+    ibmq_factory.disable_account()
+
+    return provider_to_return